+++ amanda/vm/src/binload.rs
