+++ amanda/vm/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod ama_value;
+pub mod binload;
 pub mod builtins;
-pub mod parse;
 pub mod vm;
